feat: install warp

+++ src/setup/setup-cli.bash
@@ -47,19 +47,27 @@ if ! grep -q "# Add starship to bash" ~/.bashrc; then
 # Add starship to bash
 eval "$(starship init bash)"' >>~/.bashrc
 
-    # starship preset pastel-powerline > ~/.config/starship.toml
-    cat ./src/setup/starship.toml >~/.config/starship.toml
-
 fi
 
-# Nerd fonts
-PWD=$(pwd)
-FONT_NAME="FiraCode"
-curl -LJO "https://github.com/ryanoasis/nerd-fonts/releases/download/v2.3.3/$FONT_NAME.zip"
-mkdir -p $HOME/.local/share/fonts
-mv "$FONT_NAME.zip" $HOME/.local/share/fonts/
-cd $HOME/.local/share/fonts/
-unzip "$FONT_NAME.zip"
-rm "$FONT_NAME.zip"
-fc-cache -fv
-cd $PWD
+# starship preset pastel-powerline > ~/.config/starship.toml
+cat ./src/setup/starship.toml >~/.config/starship.toml
+
+if [[ "$DESKTOP" -eq 1 ]]; then
+    # Nerd fonts
+    PWD=$(pwd)
+    FONT_NAME="FiraCode"
+    curl -LJO "https://github.com/ryanoasis/nerd-fonts/releases/download/v2.3.3/$FONT_NAME.zip"
+    mkdir -p $HOME/.local/share/fonts
+    mv "$FONT_NAME.zip" $HOME/.local/share/fonts/
+    cd $HOME/.local/share/fonts/
+    unzip "$FONT_NAME.zip"
+    rm "$FONT_NAME.zip"
+    fc-cache -fv
+    cd $PWD
+fi
+
+# Install Warp terminal
+if [[ "$DESKTOP" -eq 1 ]]; then
+    curl -LJO https://releases.warp.dev/stable/v0.2024.03.26.08.02.stable_02/warp-terminal_0.2024.03.26.08.02.stable.02_amd64.deb
+    sudo nala install ./warp-terminal_0.2024.03.26.08.02.stable.02_amd64.deb
+fi
